+++ obj/src/data/index.d.ts
@@ -14,3 +14,5 @@ export { SortParams } from './SortParams';
 export { PagingParams } from './PagingParams';
 export { DataPage } from './DataPage';
 export { FilterParams } from './FilterParams';
+export { MultiString } from './MultiString';
+export { TagsProcessor } from './TagsProcessor';